+++ tests/testthat/test_conversion.R
@@ -7,7 +7,7 @@ datw <- date2week(dat, 5)
 
 test_that("aweek prints as expected", {
 
-  expect_output(wtad <- print(datw), "aweek start: 5")
+  expect_output(wtad <- print(datw), "aweek start: Friday")
   expect_identical(wtad, datw)
 
 })
@@ -23,7 +23,7 @@ test_that("a character can be converted to a date", {
 
 test_that("aweek can be converted to character", {
 
-  expect_failure(expect_output(print(as.character(datw)), "aweek start: 5"))
+  expect_failure(expect_output(print(as.character(datw)), "aweek start: Friday"))
 
 })